ble-scanner/requirements.txt
psycopg[binary]
paho-mqtt
bleak